Debugged entities

+++ diamond-backend/diamond/src/main/java/com/diamond/diamond/entities/payments/CheckoutPayment.java
@@ -1,9 +1,9 @@
 package com.diamond.diamond.entities.payments;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 import com.diamond.diamond.entities.Customer;
-import com.diamond.diamond.entities.Payment;
 import com.diamond.diamond.entities.Vendor;
 import com.diamond.diamond.entities.VendorWallet;
 import com.diamond.diamond.types.Blockchain;
@@ -26,7 +26,7 @@ public class CheckoutPayment extends Payment {
     @Column(name="promo_codes_enabled", nullable=false)
     private Boolean enablePromoCodes;
 
-    public CheckoutPayment(double amount, Vendor vendor, Customer customer, StablecoinCurrency currency, Blockchain chain, List<VendorWallet> vendorWallets) throws Exception {
+    public CheckoutPayment(BigDecimal amount, Vendor vendor, Customer customer, StablecoinCurrency currency, Blockchain chain, List<VendorWallet> vendorWallets) throws Exception {
         super(amount, vendor, customer, currency, chain, vendorWallets);
     }
